Extend SUBSCRIPTIONDATALEN for large MQTT responses
- **Description** 
Sometimes messages with large payloads must be received from the cloud.
E.g. a list with available devices must be received. In such cases if another informations are added to 
the message too, the maximal buffer size of 100 bytes is too small. 
The message is truncated and can not be processed correctly.
The problem occurred with us with the ESP8266 device.

- **Limitations**
I have not seen any limitations on increasing the MAXBUFFERSIZE to 500 bytes.