Process MQTT messages with 1,2,3 length-bytes
- **Description** 
Long MQTT messages can not have 1 but 2, 3 ... length bytes in the header.
In this case the position of the topic length is shifted further from message ID [3] to [4], [5].
The offset for message evaluation is increased from 4 to 5, 6 ...
This must be determined automatically by evaluation of the 7th bit in length byte.
If the 7th bit in the length byte is set to 1, the next length byte is present below.


- **Limitations**
Theoretically, even up to 4 length bytes are possible, but that will not happen for this hardware.
The values for MAXBUFFERSIZE and SUBSCRIPTIONDATALEN should also be increased if the message length 
exceeds 127 bytes and there are at least 2 length bytes in the header.

- **Testing**  
I've tested these changes with an ESP8266 device and Messges longer than 127 bytes. 
This works correctly if MAXBUFFERSIZE and SUBSCRIPTIONDATALEN were also increased.