Allow connection to brokers with no authentication
Added a check condition in case the username and password are equal to "" in the .ino file.  For example, connection to test.mosquitto.org requires no authentication.  This allows ease of use for simply testing that the arduino and fona are working properly together.   Tested to function on an arduino uno, with a fona 808, sending test messages to test.mosquitto.org:1883.

Without the check condition, and the username and password set to "", the library still asserts the MQTT_CONN_USERNAMEFLAG and MQTT_CONN_PASSWORDFLAG flags, and extends the overall header by four bytes.  

Example packet:
0x10 0x10 0x0 0x4 0x4D 0x51 0x54 0x54 0x4 0xFFFFFFC2 0x1 0x2C 0x0 0x0 0x0 0x0 0x0 0x0
Corrected packet:
0x10 0xC 0x0 0x4 0x4D 0x51 0x54 0x54 0x4 0x2 0x1 0x2C 0x0 0x0